requirements.txt
flask>=2.2.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0


